- ref: combine transcript history and real time into one activity
- feat: add no connection pop up in main activity
