+++ app/src/main/java/com/example/adabv2/TranscriptActivity.java
@@ -15,47 +15,53 @@ import com.example.adabv2.Manager.ApiClient;
 import com.example.adabv2.Model.Response;
 import com.example.adabv2.Model.TranscriptHistory;
 import com.example.adabv2.Model.TranscriptRequest;
-import com.example.adabv2.databinding.ActivityTranscriptHistoryBinding;
+import com.example.adabv2.databinding.ActivityTranscriptBinding;
 
+import java.net.URISyntaxException;
 import java.util.List;
 
+import io.socket.client.IO;
+import io.socket.client.Socket;
 import retrofit2.Call;
 import retrofit2.Callback;
 
-public class TranscriptHistoryActivity extends AppCompatActivity {
+public class TranscriptActivity extends AppCompatActivity {
 
-    private ActivityTranscriptHistoryBinding binding;
+    private ActivityTranscriptBinding binding;
     private ImageView buttonBack;
-    private TextView textTranscript;
+    private TextView textRealTimeTV;
     private LinearLayout noTranscript;
     private ScrollView scrollView;
+    private Socket socket;
     private Integer sessionId;
+    private Boolean sessionHasPassed;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        binding = ActivityTranscriptHistoryBinding.inflate(getLayoutInflater());
+        binding = ActivityTranscriptBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
         init();
         getTranscriptHistory();
         buttonOnClick();
+        if(!sessionHasPassed) {
+            connectSocket();
+        }
     }
 
     private void init() {
         buttonBack = binding.buttonBack;
         TextView sessionNameTV = binding.sessionName;
-        textTranscript = binding.transcriptText;
+        textRealTimeTV = binding.textRealTime;
         scrollView = binding.scrollView;
         noTranscript = binding.noTranscriptView;
 
+        // get data from intent
         sessionId = getIntent().getIntExtra("sessionID", 0);
         String sessionName = getIntent().getStringExtra("sessionName");
+        sessionHasPassed = getIntent().getBooleanExtra("sessionHasPassed", true);
         sessionNameTV.setText(sessionName);
-    }
-
-    private void buttonOnClick() {
-        buttonBack.setOnClickListener(v -> finish());
     }
 
     private void getTranscriptHistory() {
@@ -63,6 +69,7 @@ public class TranscriptHistoryActivity extends AppCompatActivity {
         transcriptRequest.setSession_id(sessionId.toString());
 
         Call<Response<TranscriptHistory>> transcriptResponseCall = ApiClient.request().saveTranscriptHistory(transcriptRequest);
+
         transcriptResponseCall.enqueue(new Callback<Response<TranscriptHistory>>() {
             @Override
             public void onResponse(@NonNull Call<Response<TranscriptHistory>> call, @NonNull retrofit2.Response<Response<TranscriptHistory>> response) {
@@ -75,12 +82,10 @@ public class TranscriptHistoryActivity extends AppCompatActivity {
                         noTranscript.setVisibility(View.VISIBLE);
                         scrollView.setVisibility(View.INVISIBLE);
                     } else {
-                        textTranscript.setText(transcriptHistory.getMessage());
+                        textRealTimeTV.setText(transcriptHistory.getMessage());
                     }
                 }
                 else {
-                    noTranscript.setVisibility(View.VISIBLE);
-                    scrollView.setVisibility(View.INVISIBLE);
                     Log.e("Api Error", "Failed to Fetch Data");
                 }
             }
@@ -90,5 +95,31 @@ public class TranscriptHistoryActivity extends AppCompatActivity {
                 Log.wtf("responses", "Failed " + t.getLocalizedMessage());
             }
         });
+    }
+
+    private void buttonOnClick() {
+        buttonBack.setOnClickListener(v -> finish());
+    }
+
+    private void connectSocket() {
+        try {
+            socket = IO.socket("https://adab.arutala.dev/");
+        } catch (URISyntaxException e) {
+            e.printStackTrace();
+        }
+        socket.connect();
+
+        while (!socket.connected()) {
+            Log.d("Socket.io", "connecting...");
+        }
+
+        socket.emit("join_room", String.valueOf(sessionId));
+
+        socket.on("message", args ->
+            runOnUiThread(() -> {
+                textRealTimeTV.setText(args[0].toString());
+                scrollView.fullScroll(View.FOCUS_DOWN);
+            })
+        );
     }
 }
