+++ app/src/main/res/drawable/ic_no_connection.xml
@@ -1,0 +1,5 @@
+<vector android:height="100dp" android:tint="@color/blue"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="100dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M21,11l2,-2c-3.73,-3.73 -8.87,-5.15 -13.7,-4.31l2.58,2.58c3.3,-0.02 6.61,1.22 9.12,3.73zM19,13c-1.08,-1.08 -2.36,-1.85 -3.72,-2.33l3.02,3.02 0.7,-0.69zM9,17l3,3 3,-3c-1.65,-1.66 -4.34,-1.66 -6,0zM3.41,1.64L2,3.05 5.05,6.1C3.59,6.83 2.22,7.79 1,9l2,2c1.23,-1.23 2.65,-2.16 4.17,-2.78l2.24,2.24C7.79,10.89 6.27,11.74 5,13l2,2c1.35,-1.35 3.11,-2.04 4.89,-2.06l7.08,7.08 1.41,-1.41L3.41,1.64z"/>
+</vector>
