+++ app/src/main/res/layout/activity_main.xml
@@ -28,6 +28,59 @@
         android:visibility="visible" />
 
     <com.google.android.material.card.MaterialCardView
+        android:id="@+id/no_internet"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:cardBackgroundColor="@color/white"
+        app:cardElevation="5dp"
+        app:cardCornerRadius="28sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        android:visibility="invisible">
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:gravity="center"
+            android:orientation="vertical"
+            android:weightSum="5.5"
+            android:paddingVertical="25dp"
+            android:paddingHorizontal="50dp">
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="0dp"
+                android:textSize="22sp"
+                android:text="@string/noInternet"
+                android:textStyle="bold"
+                android:layout_weight="1"
+                android:gravity="center_vertical"
+                android:textColor="@color/black"/>
+
+            <ImageView
+                android:layout_width="wrap_content"
+                android:layout_height="0dp"
+                android:layout_weight="3.2"
+                android:layout_gravity="center"
+                android:src="@drawable/ic_no_connection"/>
+
+            <Button
+                android:id="@+id/button_back_to_login"
+                android:layout_width="wrap_content"
+                android:layout_height="0dp"
+                android:layout_weight="1.3"
+                android:paddingHorizontal="30dp"
+                android:layout_gravity="center"
+                android:background="@drawable/button_regis_login"
+                android:backgroundTint="@color/orange"
+                android:text="@string/backToLogin"
+                app:backgroundTint="@null"/>
+            
+        </LinearLayout>
+    </com.google.android.material.card.MaterialCardView>
+
+    <com.google.android.material.card.MaterialCardView
         android:id="@+id/progressBar"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
