+++ app/src/main/res/layout/activity_transcript.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TranscriptRealtimeActivity">
+    tools:context=".TranscriptActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
