+++ app/src/main/res/layout/fragment_register.xml
@@ -23,7 +23,7 @@
                 android:id="@+id/name"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/sesi"
+                android:text="@string/menuRegister"
                 android:textColor="@color/white"
                 android:gravity="center"
                 android:layout_gravity="center_vertical"
